refactor: Update banners grid layout based on number of banners
The banners.tsx file has been modified to update the grid layout based on the number of banners. If there are more than 3 banners, the grid will have 4 columns (lg:grid-cols-4), otherwise it will have 3 columns (lg:grid-cols-3). This change improves the visual presentation of the banners and ensures that they are displayed properly regardless of the number of banners available.