refactor: Update dataApi to handle revalidate option for Genshin data
The dataApi module has been updated to handle the revalidate option specifically for Genshin data. If the "genshin-data" tag is included in the request, the revalidate option will be set to 0, disabling revalidation of the data. This change ensures that Genshin data is always up to date and improves the performance of fetching Genshin data.