refactor: Add support for Zenless API in dataApi
This commit adds support for the Zenless API in the dataApi module. It imports the ZENLESS_API_URL constant from the constants module and includes it in the getZenlessData function. This allows fetching data from the Zenless API using the specified resource parameter. The addition of this functionality expands the data retrieval capabilities of the application.