refactor: Update resource parameter in getZenlessData function
This commit updates the resource parameter in the getZenlessData function in the disk-drives page. It changes the value from "diskDrives" to "disk-drives" to align with the API endpoint. This modification ensures that the correct data is fetched from the Zenless API and improves the accuracy of the displayed disk drives information.