refactor: Update package dependencies
Update the versions of various packages in the package.json file. This includes updating "@tailwindcss/forms" to version 0.5.8, "@tailwindcss/typography" to version 0.5.15, "@tanstack/react-table" to version 8.20.5, "nanostores" to version 0.11.3, and "next" to version 14.2.7. Also update devDependencies such as "@next/bundle-analyzer" to version 14.2.7, "@next/env" to version 14.2.7, "@types/react" to version 18.3.5, "drizzle-kit" to version 0.24.2, "eslint-config-next" to version 14.2.7, and "postcss" to version 8.4.44.